+++ src/main/java/com/block_chain/KLTN/domain/order/CreateOrderRequest.java
@@ -1,9 +1,10 @@
 package com.block_chain.KLTN.domain.order;
 
-import com.block_chain.KLTN.domain.order.order_item.OrderItemRequest;
-
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
+
+import com.block_chain.KLTN.domain.order.orderItem.OrderItemRequest;
+
 import java.util.List;
 
 public record CreateOrderRequest(
