+++ src/main/java/com/block_chain/KLTN/domain/order/OrderEntity.java
@@ -1,7 +1,7 @@
 package com.block_chain.KLTN.domain.order;
 
 import com.block_chain.KLTN.common.AbstractEntity;
-import com.block_chain.KLTN.domain.order.order_item.OrderItemEntity;
+import com.block_chain.KLTN.domain.order.orderItem.OrderItemEntity;
 import com.block_chain.KLTN.domain.organization.OrganizationEntity;
 import com.block_chain.KLTN.domain.transferObject.TransferObjectEntity;
