+++ src/main/java/com/block_chain/KLTN/domain/order/OrderResponse.java
@@ -3,7 +3,7 @@ package com.block_chain.KLTN.domain.order;
 import java.time.OffsetDateTime;
 import java.util.List;
 
-import com.block_chain.KLTN.domain.order.order_item.OrderItemResponse;
+import com.block_chain.KLTN.domain.order.orderItem.OrderItemResponse;
 
 public record OrderResponse(
     Long id,
