+++ src/main/java/com/block_chain/KLTN/domain/order/orderItem/OrderItemEntity.java
@@ -1,4 +1,4 @@
-package com.block_chain.KLTN.domain.order.order_item;
+package com.block_chain.KLTN.domain.order.orderItem;
 
 import javax.persistence.*;
